face-recognition-service/requirements.txt
Flask==2.3.2
opencv-python==4.8.0.76
dlib==19.24.2
numpy==1.25.2
mysql-connector-python==8.1.0
redis==5.0.1
